feat: move danger-zone deletion emails onto the shared templates package and delete the off-brand standalone emails.go wrapper
